+++ NEWS.md
@@ -16,6 +16,10 @@ New features(Analysis):
   1. Has no declared properties (`TypeNoPropertiesForeach`)
   2. Has properties and none are accessible. (`TypeNoAccessiblePropertiesForeach`)
   3. Has properties and some are accessible. (`TypeSuspiciousNonTraversableForeach`)
++ Add `@phan-template` and `@phan-inherits` as aliases for `@template` and `@inherits`
+
+Maintenance:
++ Minor performance improvements.
 
 Bug fixes:
 + Don't crash when parsing an invalid cast expression. Only the fallback/polyfill parsers were affected.